requirements.txt
aenum>=3.1.15
annotated-types>=0.7.0
anyio>=4.8.0
blinker>=1.9.0
certifi>=2025.1.31
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
distro>=1.9.0
Flask>=3.1.0
Flask-Cors>=5.0.0
h11>=0.14.0
httpcore>=1.0.7
httpx>=0.28.1
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.5
jiter>=0.8.2
MarkupSafe>=3.0.2
numpy>=2.2.2
openai>=1.61.0
pandas>=2.2.3
pathvalidate>=3.2.3
pydantic>=2.10.6
pydantic_core>=2.27.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-slugify>=8.0.4
pytz>=2025.1
regex>=2024.11.6
requests>=2.32.3
six>=1.17.0
sniffio>=1.3.1
text-unidecode>=1.3
tiktoken>=0.8.0
tqdm>=4.67.1
typing_extensions>=4.12.2
tzdata>=2025.1
urllib3>=2.3.0
Werkzeug>=3.1.3
